Add functionality to retrieve and display list of expenses
In this commit, a test case was added for a successful retrieval of expenses, along with corresponding updates in ExpensesController and the addition of ExpenseFixture for tests. The README was also updated to reflect this new feature of displaying a list of expenses. The UserFixtures was also updated to add a reference to a user for test cases.