Refactor expense category as Entity and update fixtures
Refactored the "category" attribute of Expense Entity from a plain string to a reference to a new Category Entity. Implemented necessary changes on the ExpensesManager, and updated test fixtures in accord with these changes. Added new file CategoryFixtures.php, and new methods in the ExpensesController and ExpensesManager for handling the category attribute.